Taking Helper classes out of the RawRead adding __init__.py
