+++ PyLTSpice/RawWrite.py
@@ -18,7 +18,8 @@ It can be used to combine RAW files generated by different Simulation Runs
 from typing import Union
 from time import strftime
 
-from PyLTSpice.RawRead import DataSet, RawRead, Trace as TraceRead
+from PyLTSpice.RawRead import RawRead
+from PyLTSpice.RawClasses import DataSet
 from numpy import array, float32, zeros
