+++ PyLTSpice/__init__.py
@@ -1,1 +1,15 @@
 # -*- coding: utf-8 -*-
+
+# import RawRead
+# import RawWrite
+# import SpiceBatch
+
+# Convenience direct imports
+from .RawRead import RawRead, SpiceReadException
+from .RawWrite import RawWrite, Trace
+from .SpiceBatch import SpiceEditor, SimCommander
+
+# Compatibility with previous code
+LTSpice_RawRead = RawRead
+LTSpice_RawWrite = RawWrite
+LTSpiceBatch = SpiceBatch
