+++ tests/AC - STEP_1.log
@@ -1,10 +1,10 @@
-Circuit: * C:\sandbox\PyLTSpice_GitHub_nunobrum\tests\AC - STEP.asc
+Circuit: * C:\sandbox\PySpice\tests\AC - STEP.asc
 
 .step r1=1000
 .step r1=10000
 
-Date: Sun Oct 09 21:56:18 2022
-Total elapsed time: 0.003 seconds.
+Date: Sun Dec 04 22:36:56 2022
+Total elapsed time: 0.008 seconds.
 
 tnom = 27
 temp = 27
