+++ tests/DC op point - STEP_1.log
@@ -1,9 +1,9 @@
-Circuit: * C:\sandbox\PyLTSpice_GitHub_nunobrum\tests\DC op point - STEP.asc
+Circuit: * C:\sandbox\PySpice\tests\DC op point - STEP.asc
 
 Direct Newton iteration for .op point succeeded.
 
-Date: Sun Oct 09 21:56:29 2022
-Total elapsed time: 0.013 seconds.
+Date: Sun Dec 04 22:37:12 2022
+Total elapsed time: 0.035 seconds.
 
 tnom = 27
 temp = 27
