+++ tests/TRAN - STEP_1.log
@@ -1,4 +1,4 @@
-Circuit: * C:\sandbox\PyLTSpice_GitHub_nunobrum\tests\TRAN - STEP.asc
+Circuit: * C:\sandbox\PySpice\tests\TRAN - STEP.asc
 
 .OP point found by inspection.
 .step vin=1 r1=1000
@@ -43,8 +43,8 @@ Measurement: t5
      4	3.93469	0.005
 
 
-Date: Sun Oct 09 21:56:33 2022
-Total elapsed time: 0.015 seconds.
+Date: Sun Dec 04 22:37:18 2022
+Total elapsed time: 0.074 seconds.
 
 tnom = 27
 temp = 27
@@ -57,6 +57,6 @@ rejected = 1
 matrix size = 3
 fillins = 0
 solver = Normal
-Matrix Compiler1: 50 bytes object code size  0.0/0.0/[0.0]
-Matrix Compiler2: off  [0.0]/0.0/0.0
+Matrix Compiler1:       3 opcodes  0.0/[0.0]/0.0
+Matrix Compiler2: 175 bytes object code size  0.1/0.1/[0.0]
